AI Inky + Clyde / TP labyrinthe
AI is now 100% done, there's a serious need to refactor some stuff but i doubt we'll have time to ever fix it so, sadly this code base will do.

Side note : The movement system needs to be redone,  to add cornering (the ability to change direction quicker than ghosts). Either that or we only let the entities move on each music beat, that could be interesting too